+++ plugins/aves_map/lib/src/style.dart
@@ -3,9 +3,6 @@ enum EntryMapStyle {
   googleNormal,
   googleHybrid,
   googleTerrain,
-  // Huawei
-  hmsNormal,
-  hmsTerrain,
   // Leaflet
   // browse providers at https://leaflet-extras.github.io/leaflet-providers/preview/
   osmHot,