+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## <a id="unreleased"></a>[Unreleased]
 
+### Removed
+
+- `huawei` app flavor
+
 ## <a id="v1.11.4"></a>[v1.11.4] - 2024-07-09
 
 ### Added